requirements.txt
asgiref==3.7.2
Django==4.2.1
django-filter==23.2
djangorestframework==3.14.0
Pillow==9.5.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.2
